fix(#391): small fixes per CR Minor threads
page_models.TableOfContents:
  Group entries by section via a dict keyed on first-appearance order
  instead of only grouping contiguous runs. Input like [A, B, A] now
  renders one A block + one B block, not three blocks.

survey/crosstab._base_respondents:
  Return float instead of int to preserve fractional weighted bases
  (1234.7 no longer silently truncates to 1234). base_note formats via
  int(round(n)) for display-only. View.base field type updated to float.

tests/test_survey_models.py:
  test_add_cluster now asserts Stack.add_cluster returns self, matching
  Cluster.add_chain's symmetric fluent-return contract.

tests/test_survey_significance.py:
  Two new tests cover the scipy-less fallback path deterministically
  via monkeypatch: one asserts the z=1.96 fallback still flags extreme
  differences at alpha=0.05, the other asserts unknown alpha (e.g. 0.02)
  raises SignificanceError rather than silently coercing.

tests/test_survey_weights.py:
  New TestApplyRimWeightsWithFakeWeightipy class injects a fake
  weightipy module into sys.modules and asserts our wrapper calls the
  documented API (scheme_from_dict with rim_params + weight_dataframe
  with weight_column). Runs in the slim CI env where weightipy isn't
  installed; catches API drift.

tests/test_survey_render.py:
  test_render_path_invokes_map_builder_and_sets_map_figure: assert
  chain_to_argument actually reaches _build_map and puts its return on
  Argument.map_figure (previous test only checked layout logic).
  test_build_map_raises_on_geo_row_mismatch: assert the new RenderError
  guard against geo_column / row_var mismatch.

+++ siege_utilities/survey/models.py
@@ -26,7 +26,7 @@ if TYPE_CHECKING:
 class View:
     """A single statistic in one cell of a cross-tabulation."""
     metric: str                  # column name / statistic label
-    base: int                    # unweighted respondent count
+    base: float                  # effective respondent count (weighted sum or raw count)
     count: float                 # weighted or raw count
     pct: Optional[float] = None  # proportion (0–1); None for non-percent views
     sig_flag: Optional[str] = None  # column letter if significantly different
